requirements.txt
flask
python-dotenv
google.generativeai